refactor(skills): swap data layer from @epicenter/filesystem to @epicenter/skills
Replace the generic Yjs virtual filesystem with the purpose-built
@epicenter/skills package. The editor no longer needs a local workspace
definition—createSkillsWorkspace() from the package handles everything.

- Replace @epicenter/filesystem + just-bash deps with @epicenter/skills + @epicenter/svelte
- Delete local workspace/ directory (definition, factory, index)
- Rewrite client.ts from 42 lines to 4 (no fs, bash, or sqliteIndex)
- Extract validateSkill to utils/validation.ts, delete types.ts
- Remove SQLite WASM and COOP/COEP config from vite.config.ts
