--- apps/skills/src/lib/workspace/index.ts
@@ -1,2 +1,0 @@
-export { skillsEditorDefinition } from './definition';
-export { createSkillsEditor } from './workspace';
